scripts/requirements.txt
google-generativeai>=0.3.0
pillow>=10.0.0
requests>=2.31.0
